BUGFIX * Decouple CHeatCloudProjectile from explotex, ie. add heatcloudtex (heatcloud in resources.tdf)   so it can be overridden separately from explotex. * Fix a few gcc warnings.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2119 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Misc/Feature.cpp
@@ -342,7 +342,7 @@ bool CFeature::Update(void)
 				float3 speed=gu->usRandVector()*0.5f;
 				speed.y+=2.0f;
 				
-				new CGeoThermSmokeProjectile(gu->usRandVector()*10 + float3(pos.x,pos.y-10,pos.z),speed,50+gu->usRandFloat()*7, this);
+				new CGeoThermSmokeProjectile(gu->usRandVector()*10 + float3(pos.x,pos.y-10,pos.z),speed,int(50+gu->usRandFloat()*7), this);
 			}
 		}
 		POP_CODE_MODE;
